Fixed merge conflict css issues
