+++ LiftOff-Project/ClientApp/src/App.js
@@ -2,7 +2,6 @@ import React, { Component } from 'react';
 import { Route, Routes } from 'react-router-dom';
 import AppRoutes from './AppRoutes';
 import { Layout } from './components/Layout';
-import css from './custom.module.css';
 
 export default class App extends Component {
   static displayName = App.name;
