+++ LiftOff-Project/ClientApp/src/components/Global.css
@@ -19,3 +19,7 @@ html {
   background-color: white;
   font-family: "HelveticaNeue-Light", "Helvetica Neue Light", "Helvetica Neue", Helvetica, Arial, "Lucida Grande", sans-serif;
 }
+
+body {
+  background-color: #871F78;
+}
