+++ LiftOff-Project/ClientApp/src/components/HomePage/SearchForm.module.css
@@ -1,0 +1,6 @@
+﻿.right {
+    position: absolute;
+    right: 0px;
+    width: 300px;
+    padding: 10px;
+}
