--- LiftOff-Project/ClientApp/src/components/SearchForm.css
@@ -1,52 +1,0 @@
-﻿.error {
-    color: red;
-    font-size: 14px;
-    line-height: 0;
-}
-
-.format {
-    text-align: center;
-    margin-top: 50px;
-    margin-left: auto;
-    margin-right: auto;
-    width: 550px;
-    border: 6px solid black;
-    border-radius: 10px;
-    display: block;
-    padding-top: 20px;
-    padding-bottom: 20px;
-    background-color: white;
-}
-
-.h1 {
-    font-weight: 800;
-    font-family: "HelveticaNeue-Light", "Helvetica Neue Light", "Helvetica Neue", Helvetica, Arial, "Lucida Grande", sans-serif;
-}
-
-.h3 {
-    font-weight: bold;
-    font-family: "HelveticaNeue-Light", "Helvetica Neue Light", "Helvetica Neue", Helvetica, Arial, "Lucida Grande", sans-serif;
-}
-
-.input {
-    font-size: 18px;
-    font-family: "HelveticaNeue-Light", "Helvetica Neue Light", "Helvetica Neue", Helvetica, Arial, "Lucida Grande", sans-serif;
-    font-weight: 300;
-    padding: 10px;
-    margin: 10px;
-    line-height: 1;
-    width: 80%;
-    border-radius: 5px;
-}
-
-.click:hover {
-    background-color: rgb(31, 172, 31);
-    border-color: rgb(31, 172, 31);
-}
-
-.click {
-    width: 30%;
-    font-weight: bold;
-    margin-top: 10px;
-    font-family: "HelveticaNeue-Light", "Helvetica Neue Light", "Helvetica Neue", Helvetica, Arial, "Lucida Grande", sans-serif;
-}
