--- LiftOff-Project/ClientApp/src/custom.module.css
@@ -1,19 +1,0 @@
-/* Provide sufficient contrast against white background */
-a {
-  color: #0366d6;
-}
-
-code {
-  color: #E01A76;
-}
-
-.btn-primary {
-  color: #fff;
-  background-color: #1b6ec2;
-  border-color: #1861ac;
-}
-
-body {
-    background-color: #871F78;
-    color: #CCCCCC;
-}
